+++ src/burlap/oomdp/core/values/DiscreteValue.java
@@ -117,7 +117,7 @@ public class DiscreteValue extends Value {
 	if (this.discVal == -1) {
 	    throw new UnsetValueException();
 	}
-	return this.discVal;
+	return (double) this.discVal;
     }
 
     @Override
